Bug 32061: Remove spans from page titles for adding/editing Z39.50/SRU servers
In another bug we added spans to ease translations in the PROCESS
block for server types. This PROCESS is used in several places,
but only creates a problem when used in the title.

The patch removes the spans using the remove TT filter avoiding
having to change the PROCESS and causing a string change.

To test:
* Go to administration > Z39.50/SRU servers
* Add a new Z39.50 and a SRU server or look at the form
* The page title in the browser tab will show the span
* Same for editing an existing Z39.50 or SRU server
* Apply the patch
* Verify the page titles are now correct

Hint: you can see the full page title when hovering on the tab.

Signed-off-by: David Nind <david@davidnind.com>
Signed-off-by: Hammat Wele <hammat@inlibro.com>
Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>
(cherry picked from commit 5dff3886b3117205a49badd03ca6f234e4531936)
Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>