Bug 32241: Add page-section to list of records in batch record modificaton (tool)
This patch modifies the batch record modification page template so that
the preview of submitted records is contained in a .page-section
container.

To test, apply the patch and go to Cataloging -> Batch record
modification.

Submit a batch of bibliographic record numbers and a batch of authority
record numbers. In both cases the page which previews the list of
submitted records in a table should be correctly styled.

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>
Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>