Bug 32319: Give header search submit button more padding
This patch tweaks the appearance of the header search submit button so
that it has a little more spacing, making it an easier click target.

The patch also gives the button a slightly different hover state in
order to make it clearer that it's a button.

To test, apply the patch and rebuild the staff interface CSS.

View various pages in the staff interface to confirm that the header
search bar looks good and works correctly at various screen sizes.

Signed-off-by: Lucas Gass <lucas@bywatersolutions.com>

Signed-off-by: Kyle M Hall <kyle@bywatersolutions.com>
Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>
(cherry picked from commit 509e828fc7309ac3cb4bba6c56bf2e4ef34b3b4a)
Signed-off-by: Jacob O'Mara <jacob.omara@ptfs-europe.com>