Bug 32352: Make xt/check_makefile.t check directories in git index
It's failing if 'node_modules' exists. We could simply skip it, but it
seems that it would be better to actually list directories that are part
of the git index only.

The question is: do we actually support test execution outside of a git
directory?

Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>
(cherry picked from commit 8c367775470fcecdda54e28270b05d92834cd33f)
Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>