Bug 32300: Add page-section to cataloguing plugins (cat)
This patch adds a "page-section" container div around the main section
of cataloging plugins which lack it. The patch also adds a "btn
btn-primary" buttons.

This patch contains indentation changes, so please ignore whitespace
when looking at the diff.

These changes are simple enough that checking the diff is probably
enough, but if you want to test a plugin in action, you go to
Administration -> MARC bibliographic framework.

 - Choose Actions -> MARC structure on the framework you want to modify
 - Choose Actions -> Edit subfields on the tag you want to modify
 - Select the tab for the subfield you want to change.
   - In the "Other options" section near the bottom, use the "Plugin"
     dropdown to select the plugin you want to associate with that
     subfield.
 - In the basic MARC editor the subfield you updated should now have an
   icon next to it for triggering the plugin. Clicking it should trigger
   a popup window which shows the updated template.

For plugins with "authorities" in the name, configure the Authorities
structure under Administration -> Authority types.

Signed-off-by: Lucas Gass <lucas@bywatersolutions.com>
Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>