Resync Compiled CSS
At some point the opac css has gone out of sync and is throwing errors
in the pre-push hook.  After discussion with Jonathan and Tomas we think
it might be something around the bootstrap fixes after our introduction
of the Vue stack.

tl;dr - Rebuild CSS to make hooks happy.

Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>