Bug 32482: Reindent holds awaiting pickup template
This patch reindents the holds awaiting pickup template so that it has
consistent indentation.

This patch also corrects a markup error that was causing a layout issue.
Incorrectly closed <div>s caused the circulation sidebar menu to
disappear.

To test, apply the patch and test all aspects of the page, including:

- Correctly functioning tabs and DataTables
- Cancel selected holds waiting
- Cancel selected holds waiting past ReservesMaxPickUpDelay
- Cancel all holds waiting past ReservesMaxPickUpDelay
- Cancel individual holds with cancellation requests (Cancellations
  of waiting holds via the OPAC. See "Default waiting hold cancellation
  policy" in Circulation and Fine rules).

Signed-off-by: David Nind <david@davidnind.com>
Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>
Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>