Bug 32605: Restore some form styling from before the redesign
This patch pulls styles from pre-22.11 staff-global.scss related to the
styling of fieldset.rows and contents. The old CSS is more verbose but
it has been better proven to hold up in a variety of situations.

The patch makes only one template change, a minor correction to the
basket groups form.

To test, apply the patch and rebuild the staff interface SCSS.

Test as many forms as possible to confirm that everything is still
correct. Forms affected are ones shown in the main body of the page, not
in sidebars.

Some good examples: Editing patrons, subscriptions, patron categories,
notices, item types.

This patch should help outstanding bugs 32513, 32543, and 32591

Signed-off-by: David Nind <david@davidnind.com>
Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>
Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>